fix: update discount methods security
And remove some unused methods
