+++ imports/plugins/core/discounts/server/index.js
@@ -2,4 +2,3 @@
 import "./i18n";
 import "./security/discounts";
 import "./publications/discounts";
-import "./methods";
