--- imports/plugins/core/discounts/server/methods/index.js
@@ -1,1 +1,0 @@
-import "./methods";
